adds comments for algebras in genus #1860
